+++ envs/cartpole.py
@@ -42,13 +42,13 @@ class ContCartPole(gym.Env):
         # bounds
         high = np.array([
             self.x_threshold * 2,
-            np.finfo(np.float32).max,
+            np.finfo(np.float128).max,
             self.theta_threshold_radians * 2,
-            np.finfo(np.float32).max])
+            np.finfo(np.float128).max])
 
         self.action_space = spaces.Box(low=-self.force_mag, high=self.force_mag,shape=(1,),
-                                       dtype=np.float32)
-        self.observation_space = spaces.Box(-high, high, dtype=np.float32)
+                                       dtype=np.float128)
+        self.observation_space = spaces.Box(-high, high, dtype=np.float128)
 
         self.seed()
         self.viewer = None
